Add validation for Settings sot_agg_query (#445)
* Add validation for Settings sot_agg_query

This adds validation to the GoldenConfigSetting clean method for
requiring sot_agg_query to have a value when the Plugin Config has
SOTAGG_ENABLED set to True. This is to address the issue where the
Golden Config home page has an SOTAGG icon with an invalid link.

* Default sot_agg_query results to a dict with an error message

The sotagg detail view raised an exception when ENABLE_SOTAGG
was set to True but the GoldenConfigSetting did not define a
sot_agg_query. This is still a possible scenario if the Config toggles
the ENABLE_SOTAGG config, so now this page will render with a
message explaining that the sot_agg_query needs to be defined.