Fix GoldenConfig list view and csv export (#468)
The GoldenConfig List View is based on the Device Model, and so Custom
Fields and Relationships were included in the table displayed for the
GoldenConfig view. The GoldenConfig view was left to use the Device
Model, but the table is updated to use the GoldenConfig Model to use
only its columns.

The GoldenConfig CSV export has been updated to export only GoldenConfig
entries that are currently in scope by all DynamicGroups associated with
any GoldenConfigSetting entry. Additionally, any device that is in
scope but does not have a GoldenConfig entry will have a placeholder
entry added.