Add total_objs_to_delete and remove the table when deleting all objects.
These changes bring this into alignment with Nautobot 2.3.11.

This also removes the self.store_table, per the TODO it should not be necessary, I was able to confirm through testing in both the unittests and manually in the UI that removing self.store_table did not make any impact to the functionality.